requirements.txt
SQLAlchemy
Flask
requests
pydantic
waitress
loguru
python-dotenv
psycopg2
asyncpg
httpx